refactor: Move areAnalyticsDisabled to analytics/areDisabled
