+++ lib/utils/analytics/areDisabled.js
@@ -1,5 +1,5 @@
 'use strict';
 
-const configUtils = require('./config');
+const configUtils = require('../config');
 
 module.exports = Boolean(process.env.SLS_TRACKING_DISABLED || configUtils.get('trackingDisabled'));
