+++ lib/utils/analytics/index.js
@@ -8,7 +8,7 @@ const BbPromise = require('bluebird');
 const pLimit = require('p-limit');
 const fetch = require('node-fetch');
 const fse = require('fs-extra');
-const areAnalyticsDisabled = require('../areAnalyticsDisabled');
+const areAnalyticsDisabled = require('./areDisabled');
 const log = require('../log/serverlessLog');
 
 const timestampWeekBefore = Date.now() - 1000 * 60 * 60 * 24 * 7;
