+++ lib/utils/analytics/index.test.js
@@ -35,7 +35,7 @@ describe('analytics', () => {
   };
   before(() => {
     ({ report, sendPending } = proxyquire('./index.js', {
-      '../areAnalyticsDisabled': false,
+      './areDisabled': false,
       'node-fetch': url => {
         usedUrl = url;
         ++pendingRequests;
